modularized and dummied cart/checkout components
