+++ client/components/index.js
@@ -17,3 +17,4 @@ export {default as ReviewForm} from './review-form'
 export {default as Checkout} from './checkout';
 export { default as AllOrders } from './all-orders';
 export { default as CartSummary } from './cart-summary';
+export { default as CartPreview } from './cart-preview';
